# server/docs/swagger.yaml
basePath: /api
definitions:
...
  DTO.CreateCourse:
    properties:
      is_principal:
        type: boolean
      sections:
        items:
          type: object
        type: array
      trimesterId:
        type: object
    required:
    - is_principal
    - sections
    - trimesterId
    type: object
...
  DTO.SubjectStat:
    properties:
      count:
        type: integer
      date:
        type: string
      difficulty:
        type: number
      grade:
        type: number
      trimester:
        type: object
      workload:
        type: number
    type: object
...
info:
  contact: {}
  description: MetroGrama API documentation.<br><br><b>Google Login:</b> <a href="/api/auth/google/login?redirect=/swagger/index.html"
    target="_blank">Login with Google</a><br>After logging in and being redirected
    back, return to this page and use protected endpoints.
  title: MetroGrama API
  version: "1.0"
paths:
...
  /course/:
    post:
      consumes:
      - application/json
      description: Creates a course record for the authenticated student
      parameters:
      - description: Course payload
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/DTO.CreateCourse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a course (student-course relationship)
      tags:
      - course
  /course/{trimesterId}:
    get:
      consumes:
      - application/json
      description: Returns the student's course sections for the given trimester.
        If is_principal=true (default) returns principal_sections with their schedules.
      parameters:
      - description: Trimester ID
        in: path
        name: trimesterId
        required: true
        type: string
      - description: If true (default) returns principal sections; otherwise (future)
          could return other section types
        in: query
        name: is_principal
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get course sections for a trimester
      tags:
      - course
...
  /notifications/ws:
    get:
      description: Opens a websocket that pushes notification events to the authenticated
        user.
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Subscribe to notifications stream
      tags:
      - notifications
  /stats/subjects/{subjectId}:
    get:
      consumes:
      - application/json
      description: Returns aggregated stats per trimester for the given subject (count,
        avg difficulty, avg grade, avg workload).
      parameters:
      - description: Subject ID (subject code)
        in: path
        name: subjectId
        required: true
        type: string
      - default: all
        description: Which students to include in the stats
        enum:
        - all
        - friends
        - friendsFriends
        in: query
        name: studentFilter
        type: string
      - description: Comma-separated list of career RecordIDs to filter by the subject's
          belonging careers
        in: query
        name: careers
        type: string
      - description: Lower bound trimester RecordID (e.g., 2425-1)
        in: query
        name: startingTrimester
        type: string
      - description: Upper bound trimester RecordID (e.g., 2425-3)
        in: query
        name: endingTrimester
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DTO.SubjectStat'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subject stats
      tags:
      - stats
...
  /student_preferences/:
    get:
      consumes:
      - application/json
      description: Returns the authenticated student's preferences
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get current student's preferences
      tags:
      - preferences
    put:
      consumes:
      - application/json
      description: Updates the authenticated student's preferences
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update current student's preferences
      tags:
      - preferences
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth
    type: apiKey
swagger: "2.0"